refactor: optimize module and export
