+++ packages/ast/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/ast",
-  "version": "0.8.9-beta.1",
+  "version": "0.8.9-beta.2",
   "description": "ESLint x React's TSESTree AST primitive utility module.",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
@@ -43,6 +43,7 @@
     "@eslint-community/eslint-utils": "4.4.0",
     "@eslint-react/tools": "workspace:*",
     "@eslint-react/types": "workspace:*",
+    "@eslint-react/shared": "workspace:*",
     "@typescript-eslint/scope-manager": "6.10.0",
     "@typescript-eslint/types": "6.10.0",
     "@typescript-eslint/utils": "6.10.0",
