+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/core",
-  "version": "0.8.9-beta.1",
+  "version": "0.8.9-beta.2",
   "description": "ESLint x React's core module for static analysis of React core API and Patterns.",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
@@ -45,6 +45,7 @@
     "@eslint-react/jsx": "workspace:*",
     "@eslint-react/tools": "workspace:*",
     "@eslint-react/types": "workspace:*",
+    "@eslint-react/shared": "workspace:*",
     "@typescript-eslint/scope-manager": "6.10.0",
     "@typescript-eslint/type-utils": "6.10.0",
     "@typescript-eslint/types": "6.10.0",
