+++ packages/core/src/api/children.ts
@@ -1,4 +1,4 @@
-import type { RuleContext } from "@eslint-react/types";
+import type { RuleContext } from "@eslint-react/shared";
 import type { TSESTree } from "@typescript-eslint/types";
 
 import { isMemberExpressionOfReactMember } from "../internal";
