+++ packages/core/src/component/component-name.ts
@@ -1,3 +1,4 @@
+import { getFunctionIdentifier, type TSESTreeFunction } from "@eslint-react/ast";
 import type { TSESTree } from "@typescript-eslint/types";
 
 export const RE_COMPONENT_NAME = /^[A-Z]/u;
@@ -8,6 +9,12 @@ export function getComponentNameFromIdentifier(node: TSESTree.Identifier | TSEST
     : node.name;
 }
 
-export function isValidReactComponentName(name: null | string | undefined): name is string {
+export function isValidComponentName(name: null | string | undefined): name is string {
   return !!name && RE_COMPONENT_NAME.test(name);
+}
+
+export function hasNoneOrValidComponentName(node: TSESTreeFunction) {
+  const id = getFunctionIdentifier(node);
+
+  return !id || isValidComponentName(id.name);
 }
