+++ packages/core/src/component/index.ts
@@ -1,5 +1,10 @@
 export * from "./component-collector";
+export * from "./component-collector-hint";
 export * from "./component-collector-legacy";
+export * from "./component-flag";
 export * from "./component-Identifier";
 export * from "./component-init-path";
+export * from "./component-kind";
 export * from "./component-name";
+export * from "./component-render-method";
+export * from "./traverse";
