+++ packages/core/src/index.ts
@@ -1,7 +1,7 @@
 export * from "./api";
 export * from "./component";
 export * from "./context";
+export * from "./helper";
 export * from "./hook";
 export * from "./internal";
 export * from "./render-prop";
-export * from "./types";
