+++ packages/eslint-plugin-react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/eslint-plugin-react",
-  "version": "0.8.9-beta.1",
+  "version": "0.8.9-beta.2",
   "description": "ESLint x React's ESLint plugin for React related rules.",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
