+++ packages/eslint-plugin/src/index.ts
@@ -3,8 +3,8 @@ import * as jsx from "@eslint-react/eslint-plugin-jsx";
 import * as namingConvention from "@eslint-react/eslint-plugin-naming-convention";
 import * as react from "@eslint-react/eslint-plugin-react";
 import * as hooks from "@eslint-react/eslint-plugin-react-hooks";
+import type { RulePreset } from "@eslint-react/shared";
 import { entries, fromEntries } from "@eslint-react/tools";
-import type { RulePreset } from "@eslint-react/types";
 import tsParser from "@typescript-eslint/parser";
 // workaround for @typescript-eslint/utils's TS2742 error.
 import type { ESLintUtils } from "@typescript-eslint/utils";
