+++ packages/jsx/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/jsx",
-  "version": "0.8.9-beta.1",
+  "version": "0.8.9-beta.2",
   "description": "ESLint x React's TSESTree AST utility module for static analysis of JSX.",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
@@ -49,12 +49,14 @@
     "@eslint-react/ast": "workspace:*",
     "@eslint-react/tools": "workspace:*",
     "@eslint-react/types": "workspace:*",
+    "@eslint-react/shared": "workspace:*",
     "micro-memoize": "4.1.2"
   },
   "bundleDependencies": [
     "@eslint-react/ast",
     "@eslint-react/tools",
     "@eslint-react/types",
+    "@eslint-react/shared",
     "micro-memoize"
   ]
 }
