+++ packages/jsx/src/pragma/is-from-pragma.ts
@@ -1,6 +1,6 @@
 import { findVariableByName, getVariablesUpToGlobal, NodeType } from "@eslint-react/ast";
+import type { RuleContext } from "@eslint-react/shared";
 import { O } from "@eslint-react/tools";
-import type { RuleContext } from "@eslint-react/types";
 import type { TSESTree } from "@typescript-eslint/types";
 import { isMatching, match } from "ts-pattern";
