+++ packages/shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/shared",
-  "version": "0.8.9-beta.1",
+  "version": "0.8.9-beta.2",
   "description": "ESLint x React's shared constants and utilities.",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
