+++ packages/shared/src/index.ts
@@ -2,4 +2,6 @@ export * from "./case-validator";
 export * from "./common";
 export * from "./constants";
 export * from "./create-rule";
+export * from "./data";
 export * from "./merge-configs";
+export * from "./types";
