+++ packages/shared/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./rule";
+export * from "./rule-category";
+export * from "./rule-name";
