+++ packages/shared/src/types/rule-category.ts
