+++ packages/shared/src/types/rule-name.ts
