+++ packages/shared/src/types/rule.ts
