+++ packages/tools/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/tools",
-  "version": "0.8.9-beta.1",
+  "version": "0.8.9-beta.2",
   "description": "ESLint x React's std library and primitives.",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
