+++ packages/types/src/index.ts
@@ -1,4 +1,1 @@
 export * from "./react-settings";
-export * from "./rule";
-export * from "./rule-category";
-export * from "./rule-name";
